+++ lib/dartastic_opentelemetry_api.dart
@@ -55,7 +55,6 @@ export 'src/api/metrics/up_down_counter.dart' hide UpDownCounterCreate;
 export 'src/api/otel_api.dart';
 // Semantics
 export 'src/api/semantics/http_header_attribute.dart';
-export 'src/api/semantics/rum.dart';
 export 'src/api/semantics/semantics_base.dart';
 export 'src/api/semantics/semconv/semconv.dart';
 // Trace